requirements.txt
matplotlib >= 3.5.0
numpy >= 1.21.1
pyserial >= 3.5
flask >=  3.0.0
flask_cors >= 4.0.0